Update Google Search preview with the latest design
