+++ packages/social-previews/src/index.ts
@@ -1,5 +1,5 @@
 export { default as FacebookPreview } from './facebook-preview';
-export { default as SearchPreview } from './search-preview';
+export * from './google-search-preview';
 export * from './twitter-preview';
 export * from './linkedin-preview';
 export * from './constants';
